finish

+++ lib/bloc/azkar/azkar_bloc.dart
@@ -15,10 +15,13 @@ class AzkarBloc extends Bloc<AzkarEvent, AzkarState> {
   @override
   AzkarState get initialState => AzkarState(list: [], pos: 0);
 
+  String queryWhere = "1";
+
   @override
   Stream<AzkarState> mapEventToState(
     AzkarEvent event,
   ) async* {
+    //Edit Or Add in Favorate
     if (event is AddOrUpdate) {
       ZekrModel zekr = await showDialog(
         context: event.context,
@@ -27,46 +30,65 @@ class AzkarBloc extends Bloc<AzkarEvent, AzkarState> {
           zekrModel: event.zekr,
         ),
       );
-      //That's if open Custom For First Time and it insert
+      //That's if open Favorate For First Time and it insert
       if (zekr != null &&
           event.context.widget.toString().contains("Animation")) {
         add(
           OpenCategory(
-              context: event.context, categoryName: CategoryNames.Custom),
+              context: event.context, categoryName: CategoryNames.Favorite),
         );
       }
     } else if (event is ChangePosition) {
       state.pos = event.pos;
     } else if (event is OpenCategory) {
+      state.title = "أذكار ${event.categoryName}";
+      queryWhere = event.categoryName == CategoryNames.Favorite
+          ? "`isFavorite`>0"
+          : "`category`='${event.categoryName}'";
+
       //Fetsh Data
-      List<ZekrModel> lst = await ZekrModel.fromDataBase(
-          where: "`category`='${event.categoryName}'");
-      yield AzkarState(list: lst, pos: 0);
-      if (lst.isEmpty && event.categoryName == CategoryNames.Custom) {
+      List<ZekrModel> lst = await ZekrModel.fromDataBase(where: queryWhere);
+
+      //if it Favorate and empty
+      if (lst.isEmpty && event.categoryName == CategoryNames.Favorite) {
         BlocProvider.of<AnimationBloc>(event.context).add(Clicked());
         add(AddOrUpdate(context: event.context));
       } else {
-        Router.navigator.pushNamed(Router.zekr);
+        yield AzkarState(list: lst, title: state.title);
+        Router.navigator.pushNamed(event.categoryName == CategoryNames.Search
+            ? Router.search
+            : Router.zekr);
       }
+    } else if (event is Search) {
+      state.title = "نتائج البحث";
+      queryWhere = event.where;
     } else if (event is Increment) {
       await state.currentZekr.increment();
     } else if (event is Reset) {
       await state.currentZekr.reset();
     } else if (event is Delete) {
-      await state.currentZekr.delete();
+      if (state.currentZekr.category == CategoryNames.Favorite) {
+        await state.currentZekr.delete();
+      } else {
+        await state.currentZekr.convertFavorate();
+      }
 //if it the last card
       if (state.list.length == 1) {
         Router.navigator.pop();
       }
 //if it the last index
-      if (state.pos == state.list.length - 1) {
+      else if (state.pos == state.list.length - 1) {
         --state.pos;
       }
     }
-    List<ZekrModel> lst = await ZekrModel.fromDataBase(
-        where: "`category`='${state.currentZekr?.category}'");
+    List<ZekrModel> lst = await ZekrModel.fromDataBase(where: queryWhere);
 
-    if (lst != null && lst.isNotEmpty)
-      yield AzkarState(list: lst, pos: state.pos);
+    if (lst != null && lst.isNotEmpty) {
+      yield AzkarState(list: lst, pos: state.pos, title: state.title);
+    } else {
+      yield AzkarState(
+          list: [ZekrModel(name: "لا توجد نتائج", about: "")],
+          title: "لا توجد نتائج");
+    }
   }
 }
